General cleanup
-cleaned up vars and improved readability

+++ src/Unlocker.h
@@ -5,4 +5,4 @@ void CrimeCheck(RE::TESObjectREFR* center, float radius, RE::Actor* act);
 
 void TryUnlock(int milliseconds, RE::TESObjectREFRPtr trg, RE::TESObjectREFRPtr act, int cond);
 
-void LockCheck(RE::TESObjectREFRPtr refPtr, RE::TESObjectREFRPtr actPtr, float pwr, int radius);
+void LockCheck(RE::TESObjectREFRPtr refPtr, RE::TESObjectREFRPtr actPtr, float pwr, float radius);
